+++ environment.yml
@@ -7,8 +7,8 @@ dependencies:
   - ocp=7.8.1
   - vtk=*=qt*
   - pyparsing>=2.1.9
-#  - sphinx=8.1.3
-#  - sphinx_rtd_theme
+  - sphinx=9.0.4
+#  - sphinx_rtd_theme=3.1.0
   - mypy
   - codecov
   - pytest
@@ -29,7 +29,6 @@ dependencies:
   - trame-vtk
   - pip
   - pip:
-    - sphinx_rtd_theme
-    - git+https://github.com/sphinx-doc/sphinx.git@de745a9
+    - sphinx_rtd_theme==3.1.0rc1
     - --editable=.
     - git+https://github.com/cadquery/black.git@cq
